+++ FakeServer/Common/Config.cs
@@ -6,5 +6,6 @@
         public const string AsyncRoute = "async";
         public const string GraphQLRoute = "graphql";
         public const string TokenRoute = "token";
+        public const string TokenLogoutRoute = "logout";
     }
 }